+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,54 +1,12 @@
 package com.techcourse.service;
 
-import com.interface21.dao.DataAccessException;
-import com.interface21.jdbc.datasource.DataSourceUtils;
-import com.techcourse.config.DataSourceConfig;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import java.sql.Connection;
-import java.sql.SQLException;
-import javax.sql.DataSource;
 
-public class UserService {
+public interface UserService {
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-    private final DataSource dataSource;
+    User findById(final long id);
 
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.dataSource = DataSourceConfig.getInstance();
-    }
+    void save(final User user);
 
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        try (Connection connection = DataSourceUtils.getConnection(dataSource)) {
-            try {
-                connection.setAutoCommit(false);
-                final var user = findById(id);
-                user.changePassword(newPassword);
-                userDao.update(user);
-                userHistoryDao.log(new UserHistory(user, createBy));
-
-                connection.commit();
-            } catch (Exception e) {
-                connection.rollback();
-                throw new DataAccessException(e);
-            } finally {
-                DataSourceUtils.releaseConnection(connection, dataSource);
-            }
-        } catch (SQLException e) {
-            throw new DataAccessException(e);
-        }
-    }
+    void changePassword(final long id, final String newPassword, final String createdBy);
 }
